Added well formed controllers

+++ car_rental/src/main/java/com/projecttuto/vehicule_rental/dto/TicketInfoDTO.java
@@ -24,7 +24,7 @@ public class TicketInfoDTO {
 
     private RepairDemandStatus status;
 
-    private float tarif;
+    private Double tarif;
 
     private String repairName;
